requirements.txt
mpi4py
pytorch
torchvision
matplotlib
numpy
scipy
wandb